Added level_up() functionality
Added level_up() functionality. It certainly could be shortened as I had to rewirte a lot of code for each case. This should do what we need to do. We should call this function each time we want them to level up, so I didn't currently add a "Would you like to level up anything else" because my thought is that we could just recall the method if the user wanted to level up again, or have it in a loop in another function. Also it is very late.